+++ tests/database/factories/PermissionFactory.php
@@ -1,0 +1,11 @@
+<?php
+
+use Faker\Generator as Faker;
+use Seat\Web\Models\Acl\Permission;
+
+$factory->define(Permission::class, function (Faker $faker) {
+
+    return [
+        'title'    => $faker->domainWord,
+    ];
+});